# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "App Store Connect Adapters",
    "description": "HTTP API for querying an App Store Connect adapters host.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:48189"
    }
  ],
  "paths": {
...
    "/api/app-store-connect/v1.0/extensions/{adapterId}": {
      "get": {
        "tags": [
          "ExtensionFeatures"
        ],
        "summary": "Gets the extension feature URIs for an adapter.",
        "operationId": "ExtensionFeatures_GetAvailableExtensions",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of adapter extension feature URIs.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/extensions/{adapterId}/descriptor": {
      "get": {
        "tags": [
          "ExtensionFeatures"
        ],
        "summary": "Gets the descriptor for the specified extension feature.",
        "operationId": "ExtensionFeatures_GetDescriptor",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "The extension feature URI.",
            "schema": {
              "type": "string",
              "format": "uri",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain a FeatureDescriptor object.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/extensions/{adapterId}/operations": {
      "get": {
        "tags": [
          "ExtensionFeatures"
        ],
        "summary": "Gets the available operations on an extension feature.",
        "operationId": "ExtensionFeatures_GetAvailableOperations",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "The extension feature URI.",
            "schema": {
              "type": "string",
              "format": "uri",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of ExtensionFeatureOperationDescriptor objects.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/extensions/{adapterId}/operations/invoke": {
      "post": {
        "tags": [
          "ExtensionFeatures"
        ],
        "summary": "Invokes an extension adapter feature.",
        "operationId": "ExtensionFeatures_InvokeExtension",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "The URI of the operation to invoke.",
            "schema": {
              "type": "string",
              "format": "uri",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "argument": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a string describing the result of the operation. Callers are responsible for correctly parsing or inferring meaning from this result.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}